Added Pagination and whether more can load
Originally I wanted to change how my database works to grab the options
separately instead of having the whole query run in succession. A major
error in my part was not making it modular incase anything need to be
added for one option only. I then decided to change it within the query
which proved to be too time consuming, and instead, I opted to just modify
and check the results post query, where if loading is needed, it will
throw a response for the option.